% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negenes-package.R
\docType{package}
\name{negenes-package}
\alias{negenes-package}
\title{negenes: Estimating the Number of Essential Genes in a Genome}
\description{
Estimating the number of essential genes in a genome on the basis of data from a random transposon mutagenesis experiment, through the use of a Gibbs sampler. Lamichhane et al. (2003) \doi{10.1073/pnas.1231432100}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/negenes}
  \item Report bugs at \url{https://github.com/kbroman/negenes/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
